Fix version detection in cases where a version isn't specified.
I have this user agent for Firefox 16 on my desktop:

Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:16.0) Gecko/20100101 Firefox/16.0

The 'Ubuntu' is specified without a version number (e.g. Ubuntu/12.04).
As a result, getVersion gets horrendously confused, because it was
expecting to find "Ubuntu/" somewhere in the string.

The fix: if we try to do a split and it doesn't result in any splitting
occurring, give up.