Not actually fast
I use this library on my logfiles.
Half of the time is spent looking up IP addresses in a on-disk database, the other half is spent in httpagentparser.
The time spent parsing the log file is marginal.

This change is obviously meant as a joke, but I suggest you do some profiling. Or I might even do some myself.